Add files via upload
create_training_set.py creates a csv of the last 500 e-mails in your inbox, accompanied by extracted features that will be used to train the supervised learning model.

train_ML_model.py is the python code that is used to train a linear regression neural network to rank the importance of my e-mails on a scale of 1-10.

model.pth is the ML model created based on my training dataset and what I find important in e-mails.

real_time_email_logging.py is the code that runs constantly. It detects incoming e-mails in real time and predicts their importance on a scale of 1/10.